# Evaluate dynamic method args before assigning block to the method body.