--- a/lib/opal/nodes/csend.rb
+++ b/lib/opal/nodes/csend.rb
@@ -6,32 +6,30 @@ module Opal
     class CSendNode < CallNode
       handle :csend
 
-      attr_reader :args_temp
-
-      def initialize(*)
-        super
-
-        @args_temp = scope.new_temp
+      # temporary variable that stores method receiver
+      def receiver_temp
+        @receiver_temp ||= scope.new_temp
       end
 
-      # Safe navigator always needs a receiver
-      # to check if it's nil
-      def needs_temporary_receiver?
-        true
+      def default_compile
+        helper :send
+
+        compile_receiver
+
+        push ", #{check_receiver_code} ? nil : $send("
+        push receiver_temp
+        compile_method_name
+        compile_arguments
+        compile_block_pass
+        wrap '(', '))'
       end
 
-      def add_method(temporary_receiver)
-        push "#{temporary_receiver} = ", receiver_fragment
-        push ", #{temporary_receiver} === nil || #{temporary_receiver} == null ? "
-        push "(#{args_temp} = [], Opal.nil_returner) : "
-        push "(#{args_temp} = [", arguments_fragment, "], #{temporary_receiver}#{method_jsid})"
-        wrap "(", ")"
+      def compile_receiver
+        push "#{receiver_temp} = ", recv(receiver_sexp)
       end
 
-      def add_invocation(temporary_receiver)
-        push ".apply("
-        push apply_call_target(temporary_receiver), ", ", args_temp
-        push ")"
+      def check_receiver_code
+        "(#{receiver_temp} === nil || #{receiver_temp} == null)"
       end
     end
   end