… spec/filters/bugs/method.rb
… spec/lib/compiler/call_spec.rb
… spec/opal/core/runtime/block_send_spec.rb
… spec/opal/core/runtime/send_spec.rb