@elia 2016-08-31
given the ever increasing noise in compiled code I'm wondering if using an inline-able function like `$rb_plus` could help without sacrificing perfs, eg:

```js
(function(Opal) {
  function $block_send(obj, method, args, block) {
    method.$$p = block;
    return method.apply(obj, args);
  }
  var self = Opal.top, $scope = Opal, $scopes = [Opal], nil = Opal.nil, $breaker = Opal.breaker, $slice = Opal.slice;

  Opal.add_stubs(['$foo']);
  return $block_send(self, self.$foo, function(){ … });
})(Opal);
```