@iliabylich 2016-09-07
PR udpated. Added two methods two `Opal` object:
+ `send_method_name` - used by a plain method call, takes a method name, fires `method_missing`
+ `send_method` - used by `super`, takes a function that represents a method body, doesn't fire `method_missing`

There are no performance changes for method calls that don't take a block or a splat.

Benchmarks for splat + block:
``` ruby
def m(v, &blk)
  v + blk.call
end

10_000_000.times { |i| m(*[i]) { i } }
```
Master:
``` sh
$ tail -n14 bm_master.js | head -n5
  return ($a = ($b = (10000000)).$times, $a.$$p = (TMP_2 = function(i){var self = TMP_2.$$s || this, $c, $d, TMP_3;
if (i == null) i = nil;
  return ($c = ($d = self).$m, $c.$$p = (TMP_3 = function(){var self = TMP_3.$$s || this;

    return i}, TMP_3.$$s = self, TMP_3.$$arity = 0, TMP_3), $c).apply($d, Opal.to_a([i]))}, TMP_2.$$s = self, TMP_2.$$arity = 1, TMP_2), $a).call($b);
```

``` sh
$ repeat 5 time node bm_master.js
node bm_master.js  16,09s user 0,02s system 100% cpu 16,009 total
node bm_master.js  16,55s user 0,02s system 100% cpu 16,479 total
node bm_master.js  15,57s user 0,04s system 100% cpu 15,511 total
node bm_master.js  14,87s user 0,03s system 100% cpu 14,811 total
node bm_master.js  14,23s user 0,04s system 100% cpu 14,176 total
```
Patch (using `Opal.send_method_name`):
``` sh
$ tail -n14 bm_patch.js | head -n5
  return $send_method_name((10000000), 'times', [], (TMP_2 = function(i){var self = TMP_2.$$s || this, TMP_3;
if (i == null) i = nil;
  return $send_method_name(self, 'm', Opal.to_a([i]), (TMP_3 = function(){var self = TMP_3.$$s || this;

    return i}, TMP_3.$$s = self, TMP_3.$$arity = 0, TMP_3))}, TMP_2.$$s = self, TMP_2.$$arity = 1, TMP_2));
```

``` sh
$ repeat 5 time node bm_patch.js
node bm_patch.js  15,25s user 0,06s system 100% cpu 15,215 total
node bm_patch.js  16,04s user 0,04s system 100% cpu 15,983 total
node bm_patch.js  16,76s user 0,04s system 100% cpu 16,692 total
node bm_patch.js  15,17s user 0,04s system 100% cpu 15,118 total
node bm_patch.js  15,27s user 0,03s system 100% cpu 15,205 total
```