@iliabylich 2016-09-07
Updated PR.
Merged `Opal.send_method_name` and `Opal.send_method` to a single `Opal.send` that may take method or method name.

Closes #1181 #1176 #1173 #1499.