@meh 2016-09-07
Do we have any performance numbers after the `send` merge? Other than that looks good to me.