@iliabylich 2016-09-07
``` sh
$ repeat 5 time node bm_master.js
node bm_master.js  14,81s user 0,04s system 100% cpu 14,748 total
node bm_master.js  15,35s user 0,04s system 100% cpu 15,298 total
node bm_master.js  14,40s user 0,04s system 100% cpu 14,343 total
node bm_master.js  15,28s user 0,04s system 100% cpu 15,234 total
node bm_master.js  14,10s user 0,05s system 100% cpu 14,057 total

$ repeat 5 time node bm_patch.js 
node bm_patch.js  14,44s user 0,05s system 100% cpu 14,394 total
node bm_patch.js  15,72s user 0,05s system 100% cpu 15,667 total
node bm_patch.js  14,41s user 0,04s system 100% cpu 14,357 total
node bm_patch.js  14,44s user 0,04s system 100% cpu 14,387 total
node bm_patch.js  14,82s user 0,05s system 100% cpu 14,775 total

$ grep '$send' bm_master.js | wc -l
3
$ grep '$send' bm_patch.js | wc -l
253
```